nvim: lsp: don't use ensure_installed for LSPs (for now)
`ensure_installed` is a bit confusing for me sometimes, like I expect
stuff listed in the table to be installed, but it doesn't care about
extra installed LSPs.
Basically saves a few seconds of clicking in Mason which LSPs I want.
For now delete it, maybe I'll need it later, so will revert.

Sometimes I switch back and force between python LSPs just as an
experiment, to test which "UX" is better, which is faster etc.
If I want to completely remove (let's say) pylsp, I need to do 2 things:
comment pylsp in ensure_installed (so it won't be installed again during
the next vim openining) and then :MasonUninstall.
This removal frees me up next time when I want to toggle LSPs, so only
need to call mason command or click wanted in mason.